refactor: update the `Navbar` component path file
