+++ src/components/RootLayout.jsx
@@ -2,7 +2,7 @@
 import PropTypes from "prop-types";
 
 import Footer from "./Footer";
-import Navbar from "./Navbar";
+import Navbar from "./navbar/Navbar";
 
 export default function RootLayout({ children }) {
   return (
