+++ src/routes/Error.jsx
@@ -2,7 +2,7 @@ import { isRouteErrorResponse, Link, useRouteError } from "react-router-dom";
 import { motion } from "framer-motion";
 
 import Footer from "../components/Footer";
-import Navbar from "../components/Navbar";
+import Navbar from "../components/navbar/Navbar";
 
 export default function Error() {
   const errorResponse = useRouteError();
